# backend/docs/openapi.json
{
  "components": {
    "schemas": {
...
      "CreatePropertyDto": {
        "properties": {
          "neighborhood": {
            "example": "Centro",
            "maxLength": 120,
            "minLength": 1,
            "type": "string"
          },
          "type": {
            "allOf": [
              {
                "$ref": "#/components/schemas/UnitType"
              }
            ]
          },
          "unitNumber": {
            "example": "101-A",
            "maxLength": 40,
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "neighborhood",
          "type",
          "unitNumber"
        ],
        "type": "object"
      },
...
      "PageMetaDto": {
        "properties": {
          "limit": {
            "example": 20,
            "maximum": 100,
            "minimum": 1,
            "type": "number"
          },
          "page": {
            "example": 1,
            "minimum": 1,
            "type": "number"
          },
          "total": {
            "example": 42,
            "minimum": 0,
            "type": "number"
          },
          "totalPages": {
            "example": 3,
            "minimum": 0,
            "type": "number"
          }
        },
        "required": [
          "page",
          "limit",
          "total",
          "totalPages"
        ],
        "type": "object"
      },
...
      "PaginatedPropertiesResponseDto": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/PropertyResponseDto"
            },
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/PageMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "type": "object"
      },
...
      "ProblemDetailsDto": {
        "properties": {
          "detail": {
            "example": "A requisição contém dados inválidos.",
            "type": "string"
          },
          "errors": {
            "description": "Demais mensagens de validação, quando houver mais de uma.",
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "instance": {
            "example": "/contracts?page=0",
            "type": "string"
          },
          "requestId": {
            "example": "0f2a9a56-b0d5-4ac5-8f1e-791e9fc92347",
            "nullable": true,
            "type": "string"
          },
          "status": {
            "example": 400,
            "maximum": 599,
            "minimum": 400,
            "type": "number"
          },
          "timestamp": {
            "example": "2026-07-12T18:30:00.000Z",
            "format": "date-time",
            "type": "string"
          },
          "title": {
            "example": "BadRequestException",
            "type": "string"
          },
          "type": {
            "example": "https://tenancy-ledger.local/problems/http-400",
            "format": "uri",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "requestId",
          "timestamp"
        ],
        "type": "object"
      },
...
      "PropertyResponseDto": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "neighborhood": {
            "example": "Centro",
            "maxLength": 120,
            "type": "string"
          },
          "type": {
            "allOf": [
              {
                "$ref": "#/components/schemas/UnitType"
              }
            ]
          },
          "unitNumber": {
            "example": "101-A",
            "maxLength": 40,
            "type": "string"
          }
        },
        "required": [
          "id",
          "neighborhood",
          "type",
          "unitNumber",
          "createdAt"
        ],
        "type": "object"
      },
...
      "UnitType": {
        "enum": [
          "KITNET",
          "ROOM",
          "APARTMENT",
          "HOUSE",
          "COMMERCIAL"
        ],
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "description": "JWT de acesso.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "API para gestão auditável de inquilinos, imóveis, contratos e cobranças.",
    "title": "Tenancy Ledger API",
    "version": "1.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/properties": {
      "get": {
        "operationId": "PropertyController_list",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedPropertiesResponseDto"
                }
              }
            },
            "description": "",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Requisição inválida.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Token ausente, inválido ou expirado.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Papel sem permissão para a operação.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Limite de requisições excedido.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Erro interno inesperado.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar imóveis",
        "tags": [
          "Imóveis"
        ]
      },
      "post": {
        "operationId": "PropertyController_create",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePropertyDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyResponseDto"
                }
              }
            },
            "description": "",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Requisição inválida.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Token ausente, inválido ou expirado.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Papel sem permissão para a operação.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Já existe uma unidade com este bairro e número.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Regra de negócio ou valor de domínio inválido.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Limite de requisições excedido.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Erro interno inesperado.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cadastrar imóvel",
        "tags": [
          "Imóveis"
        ]
      }
    },
    "/properties/{id}": {
      "get": {
        "operationId": "PropertyController_get",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyResponseDto"
                }
              }
            },
            "description": "",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Requisição inválida.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Token ausente, inválido ou expirado.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Papel sem permissão para a operação.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Unidade imobiliária não encontrada.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Limite de requisições excedido.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            },
            "description": "Erro interno inesperado.",
            "headers": {
              "X-Request-ID": {
                "description": "Identificador de correlação da requisição.",
                "schema": {
                  "maxLength": 100,
                  "minLength": 1,
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Consultar imóvel",
        "tags": [
          "Imóveis"
        ]
      }
    },
...
  },
  "servers": [],
...
}